renaming classes in new VoyageAi module
