+++ langchain4j-voyage-ai-spring-boot4-starter/src/main/java/dev/langchain4j/voyageai/spring/VoyageAiScoringModelProperties.java
@@ -2,19 +2,17 @@ package dev.langchain4j.voyageai.spring;
 
 import java.time.Duration;
 
-public class EmbeddingModelProperties {
+public class VoyageAiScoringModelProperties {
 
     private String baseUrl;
     private Duration timeout;
     private Integer maxRetries;
     private String apiKey;
     private String modelName;
-    private String inputType;
+    private Integer topK;
     private Boolean truncation;
-    private String encodingFormat;
     private Boolean logRequests;
     private Boolean logResponses;
-    private Integer maxSegmentsPerBatch;
 
     public String getBaseUrl() {
         return baseUrl;
@@ -56,12 +54,12 @@ public class EmbeddingModelProperties {
         this.modelName = modelName;
     }
 
-    public String getInputType() {
-        return inputType;
+    public Integer getTopK() {
+        return topK;
     }
 
-    public void setInputType(String inputType) {
-        this.inputType = inputType;
+    public void setTopK(Integer topK) {
+        this.topK = topK;
     }
 
     public Boolean getTruncation() {
@@ -70,14 +68,6 @@ public class EmbeddingModelProperties {
 
     public void setTruncation(Boolean truncation) {
         this.truncation = truncation;
-    }
-
-    public String getEncodingFormat() {
-        return encodingFormat;
-    }
-
-    public void setEncodingFormat(String encodingFormat) {
-        this.encodingFormat = encodingFormat;
     }
 
     public Boolean getLogRequests() {
@@ -94,13 +84,5 @@ public class EmbeddingModelProperties {
 
     public void setLogResponses(Boolean logResponses) {
         this.logResponses = logResponses;
-    }
-
-    public Integer getMaxSegmentsPerBatch() {
-        return maxSegmentsPerBatch;
-    }
-
-    public void setMaxSegmentsPerBatch(Integer maxSegmentsPerBatch) {
-        this.maxSegmentsPerBatch = maxSegmentsPerBatch;
     }
 }
