+++ langchain4j-voyage-ai-spring-boot4-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfig.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-dev.langchain4j.voyageai.spring.AutoConfig
+dev.langchain4j.voyageai.spring.VoyageAiAutoConfig
